Started Game Development
Added the main scene where the maze will be, and added two building blocks for the maze, a basic corridor and a room with a pool of mysterious fluid. Also added a prototype shadow entity and a system for making it so that it will disappear if the flashlight shines on it, and will reappear at a random location about 10 seconds later (values like the range where it can spawn, how long until it will reappear, etc. have been made into easy-to-change variables)